Move GestureMonitor to separate file
